fix: on tag, look for release of epic@YY.MM.X in $ENV/epic/spack.yaml instead of $ENV/spack.yaml
In db80abd4006130f5ee64480ac7605860ed0d006b part of the `$ENV/spack.yaml` file was split int `$ENV/spack.yaml`. We have a check in the version job that runs only for tagged releases and checks whether epic of corresponding tag is present. That check should look in the new file now, or we may not be able to tag the next release.