fix: remove invalid/redundant condition from .benchmarks:nightly and eic_xl:singularity:nightly
The condition should be valid for `eic_xl:singularity:nightly` where we define
https://github.com/eic/containers/blob/a7181e1d3102d08520627988e97da55fd703f1e0/.gitlab-ci.yml#L572-L574
however for `.benchmarks:nightly` this should not work because we don't define `BUILD_TYPE` (it is mentioned in needs:parallel:matrix:`, but that should not be exposed). Although it did appear to work for https://eicweb.phy.anl.gov/containers/eic_container/-/commit/310b8bcfd07c9076daa2c78a69d6d8a5c84865ec, that was because these conditions are not used - .nightly takes precedence. This commit "fixes" them anyway.