feat: rework .nightly skip rules, rely on BUILD_TYPE
It was discovered that .nightly takes precedence in definitions of benchmark jobs. This swithces up the approach and introduces a common condition for not building nightly containers on tags and not running nightly benchmarks on tags/outside of main branch. We need to introduce BUILD_TYPE variable, but we can't matrix over it if we want to use `needs:parallel:matrix:` until we get
https://gitlab.com/gitlab-org/gitlab/-/issues/423553